requirements.txt
Flask==3.1.2
imageio==2.37.0
keras==3.13.0
matplotlib==3.10.8
numpy==2.4.0
onnxruntime==1.23.2
pandas==2.3.3
Pillow==12.0.0
protobuf==6.33.2
scipy==1.16.3
six==1.17.0
tensorboardX==2.6.4
tensorflow==2.20.0
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
